description
This project focuses on forecasting derivative market prices in the stock market using the Facebook Prophet algorithm. The stock market is a volatile market, and predicting its prices is a challenging task. However, by selecting the appropriate factors and creating appropriate predictive models, stock price movements can be accurately forecasted. In this project, we use Yahoo Finance to acquire and analyze published stock data to create and run predictive models. The Facebook Prophet algorithm is used to forecast stock prices for a period of three months. The project employs technical, fundamental, and statistical analyses to predict the stock market, and it will be useful for traders who need to make informed decisions based on stock market trends. The scope of the project, advantages, results and discussion, limitations, and future work are all discussed in detail.